refactor: use subcommand to organize the rc toolkit
