+++ packages/rctoolkit/src/cli.rs
@@ -1,18 +1,9 @@
 use std::path::PathBuf;
 
-use clap::Parser;
-use serde::Deserialize;
+use clap::{Parser, Subcommand, Args};
 
-#[derive(Debug, Deserialize)]
-pub struct Config {
-    pub palette: PathBuf,
-    pub tileset: PathBuf,
-    pub sprites: Vec<PathBuf>,
-}
-
-#[derive(Parser)]
-#[command(author, about, version)]
-pub struct Cli {
+#[derive(Args)]
+pub struct CreateArgs {
     /// Input file path
     pub config_filepath: PathBuf,
 
@@ -23,4 +14,17 @@ pub struct Cli {
     /// Output folder path
     #[arg(long, short, default_value = "dune2.rc")]
     pub output_file: PathBuf,
+}
+
+#[derive(Subcommand)]
+pub enum Commands {
+    Create(CreateArgs),
+}
+
+#[derive(Parser)]
+#[command(author, about, version)]
+#[command(propagate_version = true)]
+pub struct Cli {
+    #[command(subcommand)]
+    pub command: Commands,
 }
